Change getRExecutable to getRCMD, now returns R executable path with CMD appended in the right way for Unix and Windows.
Modify buildPackage and checkPackage in hopes of working in a platform
independent way.

Modify behavior of buildPrelimRepos to not build any packages up
front.  Just create a PACKAGES file from the manifest (rpacks.info2).
Changed installBuiltPackages to detect repositories with 'file:' in
their URL and to try to build the package before installing.  Note
that this code is fragile in that it depends heavily on current
working directory and we are currently doing this opperation on *all*
repositories with file: URL scheme.

Add platform detection for selecting CRAN mirror URL.  

Break out reposList creation and option munging into their own
functions.



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@10980 bc3139a8-67e5-0310-9ffc-ced21a209358