Add downloadPackage function.
Add "yesterdayURL" arg.  Allows specification of a repository in which we
search for packages that failed R CMD check and didn't make it to the newly
generated repos.  If package is found in repos at yesterdayURL, we download and
copy it over so that it will be in the repository that we build.



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@10986 bc3139a8-67e5-0310-9ffc-ced21a209358