Make reposTools faster on startup...
Remove code in .First.lib that referenced
getOption("reposTools")$autosync.  AFAICT, this option is never set by
reposTools and I don't know who else would be setting it.  This var
was set to determine if a call to syncLocalLibList() should be made
and that is part of what makes reposTools so slow on load.

git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@12200 bc3139a8-67e5-0310-9ffc-ced21a209358