Validate terms listed in biocViews DESCRIPTION field against known vocabulary.
We do the comparison normalized to lowercase and replace found terms with the 
term as found in the vocabulary (so that MixedCaseTerms will always be the same).

A warning is produced to list terms not found in the vocabulary.

This means that we cannot have terms that differ only in case.

This enhancement fixes a the check ERROR due to the Resourcerer package
listing "microarray" in its biocViews field.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@14283 bc3139a8-67e5-0310-9ffc-ced21a209358