Use open="r" when calling file() in all examples
If a given file is not found using system.file(), the return value is
"".  When file() is called with filename "", it opens tempfile 
for rw access.  If open="r" is specified, you get an error.  This 
is much better as the tempfile leads to very strange errors from 
other packages such as XML.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@19528 bc3139a8-67e5-0310-9ffc-ced21a209358