Try to fix dispatch bug in summary/htmlReport
The S4 dispatch system was failing to find the summary method
for result objects when called from the htmlReport method.

Somehow the summary method was getting set on a generic defined
in multtest (which does not have a NAMESPACE) and that wasn't
accessible from inside package code (even though selectMethod 
shows it to be there).

This patch attempts to solve this by calling setGeneric("summary")
in the GOstats package code.  Seems to fix it, but may cause
problems for other summary methods in other packages???

Also fixing generic for htmlReport using the signature arg in
setGeneric to have named args that aren't part of dispatch.




git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@19788 bc3139a8-67e5-0310-9ffc-ced21a209358