Added write_SYMBOLS and call it from genReposControlFiles
write_SYMBOLS writes a SYMBOLS file which is DCF formatted and
contains the exported symbols for each package that has a name
space.  The function can operate on a CRAN-style repos and process
tarballs in src/contrib, or it can operate on source package dirs
directly.

The purpose of this file is to allow easy indexing for webapp
package/symbol search.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@19966 bc3139a8-67e5-0310-9ffc-ced21a209358