Ugly hack so it builds.  Copy coercion code from R/src/coerce.h
Since FooFromBar functions became private in R, we reimplement
them in coerce.c locally.  This is a bummer.  We should
probably rename or get the attribute_hidden stuff to work 
properly.  Since these functions are called from a few places,
we can't just make a static definition.

Note: this patch allows the package to compile and install,
but it is currently not passing check.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@20172 bc3139a8-67e5-0310-9ffc-ced21a209358