I changed NAMESPACE and DESCRIPTION to reflect new classes for norm2Filter and filterTree
I added the file R/perry.r that has the definition of the method for applyFilter for norm2Filter. Note that this doesn't work right now because norm2Fit from prada tries to create an object of class gate that is incompatible with our virtual class gate. This would be fixed if we switched to filter.

I added the file example1.r in the directory examples so that we could keep a running example that could be gradually converted to the vignette.

Earlier I added classes norm2Filter and filterTree to AllClasses.R

git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@20667 bc3139a8-67e5-0310-9ffc-ced21a209358