Fixed the numEdges bug: the package used to install a slot in the Ragraph object with the name numEdges. This slot was not part of the class definition (but could still be access by object@numEdges - rather interesting). In fact the slot was not needed since agnedges(* Ragraph) yields the number of edges.
Fixed the 4 R_scalar*** functions from utils.h/Rgraphviz.c. The name seemed to imply they were part of
R's API, which they were not. R_scalarInteger, R_scalarReal was replaced by Rf_ScalarInteger,
Rf_ScalarReal (the function definitions were the same). R_scalarLogical which build a single logical
from an Rbool was renamed Rgraphviz_ScalarLogicalFromRbool, the existing function from Rinternals.h uses
an int instead of Rbool, so it need a separate function. The last function R_scalarString was renamed
Rgraphviz_ScalarStringOrNull, this function is actually almost the same as mkString from R's API, but in
Rgraphviz we sometimes pass a pointer to NULL to the function - that will crash mkString. I have redone
the function slightly to test for pointer to null and then calling mkString if not true.



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@21550 bc3139a8-67e5-0310-9ffc-ced21a209358