Change default for summary(GOHyperGResult) to use htmlLinks=FALSE
When used at the command line, htmlLinks=FALSE should be the default
since it is much more readable.  Had to add our own htmlReport method
so that it can get turned on by default for an HTML report.
I think this uncovered an S4 methods bug, where it should be enough
to call callNextMethod with no args, but it doesn't work.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@21848 bc3139a8-67e5-0310-9ffc-ced21a209358