ugraph is now a method with a new (faster) implementation
The old version is available as ugraphOld for a brief testing period.
The new implementation is a method defined for graph objects.  It
requires that edgeMatrix be defined for the specific graph subclass
passed in.

Currently, the return value is always an undirected graphNEL.  It
would be nice to be able to return an undirected graph in the same
representation as that which was used as input.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@22357 bc3139a8-67e5-0310-9ffc-ced21a209358