Fix graphNEL => graphAM, remove rownames from adjMat
The optimized version was resulting in a graphAM instance with
rownames and colnames on the adjMat.  The graphAM initialize method
removes rownames so that node names are sotred once (as colnames).
This patch takes the extra step of removing the rownames so that
objects will be the same from either code path.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@22362 bc3139a8-67e5-0310-9ffc-ced21a209358