I have made some changes to the configure.in script and bumped the version number.
The changes were related to problems with user-specified graphviz directory.
(1) just specifying --with-graphviz resulted in
  GRAPHVIZ_DIR="yes"
instead of
  GRAPHVIZ_DIR="/usr/local"
This is now fixed properly, unlike what I tried a few days ago.
(2) specifying --with-graphviz=$SOMETHING implicetly assumed Graphviz 2.4 or greater
as it used
  -lgvc
as part of the PKG_LIBS, but -lgvc does not work on older versions of Graphviz. I now
set the PKG_LIBS variable depending on Graphviz version. Actually I am not 100% sure
when -lgvc was introduced - I _think_ it was with Graphviz 2.3 but I don't _know_. Perhaps
a better way is to check for dotneato-config and use its output (remember it is gone
from newer versions), but I do not have time to do this today.

I have also added various configure messages so we can see what actually happens
(this is also related to a recent conversation on BioC). The output is slightly
more ugly, but we get more info in case of user problems.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@23940 bc3139a8-67e5-0310-9ffc-ced21a209358