Put ugly hack for finding summary generic into a helper function for use in GOstats
The issue is that we don't know where the summary generic will end up and so
we have to find it ourselves for now.  This should be removed as soon as a cleaner
solution is available.  Since GOstats was having the same problem, we re-use the
ugly hack so that it is easier to remove later.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@24140 bc3139a8-67e5-0310-9ffc-ced21a209358