Rename class Untyped => NullIdentifier
I don't think this class will be useful other than as a null object
placeholder.  I don't like Untyped as it is too generic.  Since the
other subclasses have "Identifier" as part of the name, I think this
one should as well.  I'm not too worried about conflict with the NULL
database.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@24786 bc3139a8-67e5-0310-9ffc-ced21a209358