Add getGoToEntrezMap_db, big speed up for hyperGTest "over"
This patch adds DB-specific code (with custom SQL, unfortunately) that
speeds up the categoryToEntrezBuilder step of hyperGTest for
GOHyperGParams when doing a test for over representation.  The under
representation case is special in that we need to keep the GO IDs that
don't have any of the selected Entrez IDs annotated at them and so
that is left as a TODO.  We also need to add DB-specific code for
the other gene set collections (KEGG, PFAM, etc).  Also, the method of
dispatching on DB-based or not ann pkgs is not ideal.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@26603 bc3139a8-67e5-0310-9ffc-ced21a209358