Improve error message for subListExtract edge case (0-length input)
When given list() as input, subListExtract will return list(), unless
simplify=TRUE, in which case an error is raised.  The reasoning is
that when the user asks to simplify, they expect a non-list to be
returned.  We can't determine what sort of vector they expect so an
error is raised.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@26758 bc3139a8-67e5-0310-9ffc-ced21a209358