Wrapper functions for chip packages
This update adds wrapper functions for chip packages along with man
pages.  This code still needs to be much more thoroughly tested and
there are still problems here with support for Yeast and Arabidopsis
which both still need to have the portion of their code that maps data
set up for the mapping of non-affy packages.  But at least now it
should be possible to call a function, and have it make a database,
and then wrap that up into a chip based package.  Organism based
packages are more monolithic and so are not really expected to need a
wrapper package.  Although the code to make the sqlite files and the
packages for these is already here.



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@29732 bc3139a8-67e5-0310-9ffc-ced21a209358