Cleans up the parameter names and input for SQLForge.
The param "ll" has been changed to "eg" for SQLForge since they are
not called locus link IDs anymore (so why name parameters after a name
people are forgetting in new code?).  Also there is an issue that can
crop up with refseq IDs if the user is not paying attention and leaves
on the version numbers to these IDs.  For Refseq IDs, only the
accession part of the ID is actually matched to a gene, and some
vendors give IDs that do have the version number attached while others
trim them off.  Therefore our DBs always deal with these IDs with the
extraneous stuff removed.  However since none of the databases use
these IDs, it also becomes impossible for any database to match on
them if the version numbers have not already been not clipped off by
the user (when the IDs have been provided by a vendor who leaves them
on), so there is now some "auto-cleanup" that SQLForge will do if the
user feeds refseq IDs as a primary ID that are still not trimmed of
their version numbers.



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@30272 bc3139a8-67e5-0310-9ffc-ced21a209358