Changes to enable AnnBuilder to work with GO.db instead of GO.
GO is going away, so these changes are needed to depend on GO.db
instead.  To complete the transformation with a minimal rewriting of
AnnbBuilder, it was also necessary to include some functions from
AnnotationDbi so that the GO environment GOTERMS could be made into a
real environment.  This was all done so that we can keep ABPkgBuilder
alive just a little bit longer for those times when someone comes
looking for annotation packages for "platypus", or any other unusual
types of model organisms.



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@32283 bc3139a8-67e5-0310-9ffc-ced21a209358