Final changes for Worm, Pig & Chicken, adds new info. to most pkgs.
This update adds support for Worms, pigs and chicken.  It also adds
suport for uniprot IDs to most packages, ZFIN IDs to zebrafish
packages, and Wormbase IDs to worm packages.  Finally, it also adds a
CHRLOCEND mapping to most packages which will allow the inclusion of
gene ending locations into most annotation packages in addition to the
usual CHRLOC mapping that already provides the start positions.



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@33781 bc3139a8-67e5-0310-9ffc-ced21a209358