Fixes a bug when interacting with the org.Sc.sgd pkg.
The org.Sc.sgd.db package is necessarily a strange beast.  It is an
organism package, (not a chip based package) but it is also based on
the sgd ORFs instead of probes.  So its not a typical org package OR
a typical yeast chip package.  This patch adresses this problem so
that this package should hopefulyl work better with GOstats etc.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@34995 bc3139a8-67e5-0310-9ffc-ced21a209358