Add the "ACtree2" container.
This container uses a more compact representation of the Aho-Corasick tree
that results from preprocessing a Trusted Band. It is an attempt to address
the problem of the size of the "ACtree" container currently in use in PDict
objects. With this new representation, a node can occupy either 8 bytes (2
ints) or 28 bytes (7 ints) in memory, depending on whether it's "extended"
(i.e. it has >= 2 links) or not (i.e. it has 0 or 1 link). Some testing with
real data tends to show that less than 10% of the nodes are typically extended
right after preprocessing (i.e. before any use of the PDict object), and that
less than 30% or 40% get extended during the typical life of the PDict object
(e.g. after it has been used on a full genome). This should reduce the size of
PDict objects by at least a factor of 2 with no major impact on performance
(no more than 10%, but it could be much less, this still needs to be verified).
This reduction in size plus some specificities of the internal representation
of ACtree2 objects allows preprocessing of Trusted Bands made of > 60 millions
35-mers (on machines with enough memory).


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@35872 bc3139a8-67e5-0310-9ffc-ced21a209358