Updated biexponential transform so that it now calculates the solution of a generic biexponential function S(x,a,b,c,d,f) = ae^(bx) -ce^(-dx) +f  instead of the logicle function it was calculating earlier.
Updated Logicle transformation: The implementation of the logicle
transform
was found to be different from that described by Parks et al. A new C
function 
was implemented for the logicle transformation reflecting the method
described 
in the paper. Additionally, the parameter argument f passed to  the
logicle 
function was missing a multiplying factor a.



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@38805 bc3139a8-67e5-0310-9ffc-ced21a209358