Fixed bug in flowPlex construction. Checks the number of instances as well as the sample names of each instance for consistency.
Added excludeTime (default TRUE) to the model constructor so that if the user
want to consider all of the parameters in an fcs file, the constructor will
attempt to exclude the parameter containing time, if it is found.

Added 'bin_cols' to model plotting, allowing the user to supply colors to the 
plotted bins, for model plotting.

Removed '<' symbol from the 'qc' and 'fancy plot color legend.

Removed references in the make system, (configure.ac) to BLAS and the gsl
library.

Removed aclocal.m4 file which is no longer needed.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@40339 bc3139a8-67e5-0310-9ffc-ced21a209358