Bug fix for missing keys.
This fixes a problem introduced when the maps were enabled for
supporting probes that map to multiple targets.  It basically fixes
Lkeys() and Llength() so that they return the correct (complete) key
sets on maps regardless of which mappings are "exposed" or "masked"
depending on their status as being probes that match multiple targets
or not.  The Lkeys() and Llength() methods should always give the FULL
set of probes even when some of the values for those probes have been
masked by a filter.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@41163 bc3139a8-67e5-0310-9ffc-ced21a209358