Adds an early stage vignette.
This vignette was originally going to go into the rtracklayer package
to explain its utilities.  But the idea was floated that perhaps we
need to go well beyond that and so this is being placed here with the
idea that it will eventually grow into explaining higher level
annotation features that will be created for GenomicFeatures.  



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@42102 bc3139a8-67e5-0310-9ffc-ced21a209358