Made most of the vignette read from disk rather than web
The vignette build process is WAY too slow, so I have not gotten a build on windows for several weeks, it appears.  Unfortunately, that means using files and sort of defeats the purpose of the vignette to show functionality of web data retrieval, but at least the data structures will be there.

git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@43229 bc3139a8-67e5-0310-9ffc-ced21a209358