Fix handling of PubmedArticle XML with many PMID children
A possibly new pubmed article XML format contains child nodes with a
PMID element related to CommentsCorrections.  getPMInfo was raising an
error because it was capturing many PMIDs for a given article instead
of one.

Also, avoid warning in if() by using inherits instead of class.  And
use seq_along instead of 1:length(...) to avoid running a loop if no
results are found.

git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@43718 bc3139a8-67e5-0310-9ffc-ced21a209358