CNSet now extends SnpSet directly
Reason for extending SnpSet directly (as opposed to SnpSuperSet).
I can define an initialization method for AlleleSet that takes
assayData as an argument rather than the AD elements.  This works
fine.  However, any attempt to extend the AlleleSet class is
problematic b/c of the setValidity defined for this class.  I'm not
entirely clear why it fails.  Forgive the lack of a reproducible
example here -- I'm just going to summarize:

If I comment out the setValidity check for AlleleSet, I can create the
following CNSet object

cnSet <- new("CNSet", <appropriate args>)
ls(assayData(cnSet))
alleleA, alleleB, call, callProbability

If I define a setValidity method for CNSet that checks for the above
assay data elements, the cnSet object is valid.  This SHOULD also be
valid for AlleleSet.  But if I source back in the setValidity method
for AlleleSet, and do

setValidity(cnSet)

it fails claiming that assay data elements alleleA, alleleB or
senseAlleleA, ... are not available.  Indeed, if I put a browser in
the setValidity for AlleleSet I can see that the object has no assay
data elements.  I know that setValidity is passed up the chain
(something like callNextMethod), but I'm not clear why the assay data
elements would not be propagated.  While convenient, setValidity and
callNextMethod functions difficult to debug.

git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@48786 bc3139a8-67e5-0310-9ffc-ced21a209358