Tweaks to get rid of warnings.
Made minor tweaks (when it was obvious what to do) to the manual
pages.  Dropped superfluous DESCRIPTION and NAMESPACE files from /R
(these can be dug out of svn if they are really needed, but an R
package can only make use of the ones in the base directory.  If they
are desired to have as a record, then they should be moved to
inst/extdata instead of /R.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@50006 bc3139a8-67e5-0310-9ffc-ced21a209358