Adjustments to KCsmart 2.9.1:
- Introduced KernSmooth package for kernel convolution

This introduces a local extimation of the density and convolution via fft. ~160 times faster,
but the algorithm has trouble in low density areas. Therefore we set low-density areas to
NA. There are more low density areas if you use a smaller kernel width. The result of the 
KernSmooth algorithm gives reliable (~.99 correlation) results for datasets with on average
5 probes/sigma.
Two causes for the difference between the new and the old version are:
1. NAs in the sampling grid are different, since we enforce low-density areas to be NA.
2. The sampling grid is different, as it is slightly adjested to accomodate the Fourier Transform.

introduced an 'old' boolean flag to the calcSpm and calcSpmCollection functions so users can still run
the old method if they want to reproduce old results.

- Removed the mirroring of probes in every iteration of the permutation

Currently the mirroring of the probes takes ~15 sec. This was repeated every iteration of
the loop. I removed the mirroring to once before the loop. In the loop itself the mirrored
probes are now drawn from the permuted values, instead of the permuted values themselves
being mirrored at the boundaries. This should not make any difference.

- Permute the data before the separation in positive and negative values

The permutation was done on both the positive and the negative data. As the permutation
took ~9 sec each this was a large time bottleneck. Now I permute and only split afterwards.
Furthermore I rewrote the permutation to a for loop instead of an apply function, this gave
a speed advantage in a 26 sample dataset (385k probes) of factor 5.

- Check for a correct class vector before doing all the samples (in comparative section)

An error was raised if the class vector was not correct but only after the kernel smoothing 
of all samples. This check is now done at the beginning to prevent unnessesary waiting on an analysis
that is going to raise an error anyway



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@52592 bc3139a8-67e5-0310-9ffc-ced21a209358