createMGsaGoSets is now exported. This function is useful for people working with IDs that are different from those in the GAF file (example Ensembl transcript ID vs Uniprot ID in human)
Append set annotations to set results when mgsa() is called on an MgsaSets. This gives more verbose output than before.
To do this the following changes have been made:
mgsa-methods: append set annotations when called wit an MgsaSets
MgsaResults: dotted table for the results
dottedTable: only add dots on factor and characters so that numbers are still completely displayed




git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@57273 bc3139a8-67e5-0310-9ffc-ced21a209358