Add function probabilityOutlier and add argument 'k' to cnEmission.
probabilityOutlier estimates the mixture probabilities for the difference of the point estimates and a median-smoothed estimate. The default for the median smoother is k=3.  The idea is that if the point estimates are much different than the median smoothed values, its more likely to be an outlier.  plot delta versus gamma.  The motivation is that zero values were causing us to jump to the altered state. But zeros occur by truncation of the cn estimates due to small phi, etc.

git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@58110 bc3139a8-67e5-0310-9ffc-ced21a209358