minor tweak to try and speed up processing of GFF files without indices.
Some GFF files don't have data to indicate exon rank.  When that
happens we can try to infer it.  This change should speed this process
up a bit.  Though really, I wonder if we support it at all since I
think it's a bad idea for people to do this.  That is, they should
really start with better files that don't require them to infer
anything (hence the warning when this feature is used).


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@72302 bc3139a8-67e5-0310-9ffc-ced21a209358