style and sort fixes
Unfortunately, this commit impacts 2 different parts of ReportingTools



1)Sorting
New functions were written in the plugins.js file. 
These changes required new classes to be used when converting from a DF to the HTML.
The new classes have "-robust" after them to distinguish them from the earlier sort methods.
The newer sort methods appear in the objectToHTML-methods and data.frame-methods

2)Styling
Twitter bootstrap is now used to style content.
All the bootstrap files have been added to the repository
Also, in openPage a number of changes have been made to 
        load the js and css files
        write the correct html classes in the html
There are also some dataTable and reprise specific css that are included




git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@74034 bc3139a8-67e5-0310-9ffc-ced21a209358