edgeR 3.1.5
The default method of plotMDS() for DGEList objects has changed.  The
default method is now to convert to moderated logCPM values and call
the plotMDS() method for numeric matrices.  The old method based on
BCV-distances is available by setting method="BCV".

glmQLFTest() can now take either a DGEList object and design matrix
or a fitted glm object.  It can also now produce a plot of the
residual deviances and shrunk deviances versus logCPM.

git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@74132 bc3139a8-67e5-0310-9ffc-ced21a209358