When the client asks for a file, we now fetch it from S3. This is slower than it used to be, because first we fetch the file from S3, then we send it to the client.
However, we will now update the client so that it fetches files directly from
S3, which will not incur this performance penalty. 

This means we can disconnect the 1TB EBS volume holding all our files (and stop paying money for it) and we can store much more than 1TB in S3 (however much
we're willing to pay for).

Still need to update recipe code so that it puts source file and derived file
in S3 when done.



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@83465 bc3139a8-67e5-0310-9ffc-ced21a209358