Change if() to while() in consensusRegion
consensusRegion: Keep halving the threshold until some counts are
beneath the threshold.

consensusRegion checks whether all counts are below a certain threshold.
If so, the threshold is divided in half.  Occasionally, this halved
threshold is still too high and all counts will remain below the
threshold. Essentially, this causes a zero-length vector to be assigned
to the variable index.start. i.e.,

index.start <- min(which(cnt > threshold)

which later results in an indexing error

git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@109929 bc3139a8-67e5-0310-9ffc-ced21a209358