Remove "nobj" method for Grouping objects added at commit 113624. It was doing:
  length(unique(unlist(x, use.names=FALSE)))

which is based on the incorrect assumption that each object in the original
collection of objects belongs to at least one group. Doing:

  max(unlist(x, use.names=FALSE))

would have been a slightly more efficient alternative but is also wrong.
There is just no way to compute the nobj() of a Grouping object in general.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@113742 bc3139a8-67e5-0310-9ffc-ced21a209358