v1.5.12 -- major version change, added DESeq2Report()
This version now allows you to set the ggplot2 theme for the plots made with
the ggplot2 package. It also has better control on rounding digits for
numerical variables that are displayed in the interactive tables. But the most
important change is the introduction of the DESeq2Report() function. It is
specifically tailored for DESeq2 results and at a minimum requires a
DESeqDataSet object with the results from a differential expression analysis
as created by DESeq2::DESeq(). See the example for more information.

git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@115539 bc3139a8-67e5-0310-9ffc-ced21a209358