Bugfix: bad search interval limits in samplesize()
Bad heuristic guessing at the search interval for the test statistic cutoff for a given top percentage of features could crash the uniroot() function in samplesize(). The default limits have now be set to fixed values suitable for the examples in the samplesize documentation, and the guessing at the search interval in other situations has been delegated to uniroot() via argument extendInt.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks@125237 bc3139a8-67e5-0310-9ffc-ced21a209358